# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Free Roulette Spin System API
  version: 1.0.0
  description: |
    ## Overview
    
    The Free Roulette Spin System API provides endpoints for managing user rewards, 
    roulette gameplay, and letter collection mechanics in a casino platform.
    
    ## Features
    
    - **Mission Management**: Earn and claim free spins through deposit-based missions
    - **Roulette Gameplay**: Spin the roulette wheel to win cash and letters
    - **Letter Collection**: Collect letters to form words and claim collection bonuses
    - **Deposit Processing**: Process deposits and evaluate mission eligibility
    - **Transaction Tracking**: Complete audit trail of all balance changes
    
    ## Authentication
    
    All endpoints require a valid User ID in the `X-User-Id` header. This ID should 
    correspond to a user in the main casino system.
    
    ## Error Handling
    
    The API uses standard HTTP status codes and returns structured error responses 
    with detailed error information including error codes and user-friendly messages.
  contact:
    name: Casino Development Team
    email: dev@casino.com
    url: https://casino.com
  license:
    name: Proprietary
    url: https://casino.com/license
servers:
  - url: http://localhost:8080/api
    description: Development server - Local development environment
  - url: https://staging-api.casino.com/roulette
    description: Staging server - Testing environment
...
security:
  - UserIdHeader: []
paths:
...
  /letters/collection:
    get:
      tags:
        - Letters
      summary: Get user's letter collection
      description: Retrieve all letters collected by the user and their current counts.
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      responses:
        '200':
          description: Successfully retrieved letter collection
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LetterCollectionDTO'
              examples:
                letter_collection:
                  summary: Letter collection
                  value:
                    - letter: "H"
                      count: 2
                    - letter: "A"
                      count: 1
                    - letter: "P"
                      count: 3
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /letters/words:
    get:
      tags:
        - Letters
      summary: Get available words
      description: Retrieve all available words for collection bonuses with user-specific claim status.
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      responses:
        '200':
          description: Successfully retrieved available words
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LetterWordDTO'
              examples:
                available_words:
                  summary: Available words
                  value:
                    - id: 1
                      word: "HAPPY"
                      requiredLetters:
                        H: 1
                        A: 1
                        P: 2
                        Y: 1
                      rewardAmount: 10.00
                      canClaim: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /letters/words/{wordId}/claim:
    post:
      tags:
        - Letters
      summary: Claim word bonus
      description: Claim a bonus for completing a word collection. Deducts required letters from user's collection.
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: wordId
          in: path
          required: true
          description: Word ID to claim
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Word bonus claimed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                success:
                  summary: Success response
                  value:
                    success: true
                    message: "Word bonus claimed successfully"
                    wordId: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          description: Insufficient letters for word
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                insufficient_letters:
                  summary: Not enough letters
                  value:
                    success: false
                    error: "INSUFFICIENT_LETTERS"
                    message: "You don't have enough letters to claim this word"
                    timestamp: "2024-01-15T10:30:00Z"
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    UserIdHeader:
      type: apiKey
      in: header
      name: X-User-Id
      description: User ID header for authentication. Must be a valid positive integer.
  parameters:
    UserIdHeader:
      name: X-User-Id
      in: header
      required: true
      description: User ID
      schema:
        type: integer
        format: int64
        minimum: 1
        example: 12345
  schemas:
...
    LetterCollectionDTO:
      type: object
      required:
        - letter
        - count
      properties:
        letter:
          type: string
          pattern: "^[A-Z]$"
          description: Single uppercase letter
          example: "H"
        count:
          type: integer
          minimum: 0
          description: Number of this letter collected
          example: 2
    LetterWordDTO:
      type: object
      required:
        - id
        - word
        - requiredLetters
        - rewardAmount
        - canClaim
      properties:
        id:
          type: integer
          format: int64
          description: Word ID
          example: 1
        word:
          type: string
          description: The word that can be formed
          example: "HAPPY"
        requiredLetters:
          type: object
          additionalProperties:
            type: integer
            minimum: 1
          description: Map of letters to required counts
          example:
            H: 1
            A: 1
            P: 2
            Y: 1
        rewardAmount:
          type: number
          format: decimal
          minimum: 0
          description: Cash reward for completing this word
          example: 10.00
        canClaim:
          type: boolean
          description: Whether the user can claim this word bonus
          example: true
...
    SuccessResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - message
        - timestamp
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error code
          example: "INSUFFICIENT_SPINS"
        message:
          type: string
          description: Human-readable error message
          example: "You don't have enough spins available"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-15T10:30:00Z"
  responses:
    BadRequest:
      description: Bad Request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_user_id:
              summary: Invalid User ID
              value:
                success: false
                error: "INVALID_USER_ID"
                message: "User ID must be a positive integer"
                timestamp: "2024-01-15T10:30:00Z"
...
    InternalServerError:
      description: Internal Server Error - Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            server_error:
              summary: Internal server error
              value:
                success: false
                error: "INTERNAL_ERROR"
                message: "An unexpected error occurred. Please try again later."
                timestamp: "2024-01-15T10:30:00Z"
...